remove indexing op-inl.h
